Set specific JAR name in Dockerfile
The Dockerfile has been updated to use a specified JAR name stored in a new environment variable. Prior to this, the file was searching for any .jar files in the build libs directory, potentially leading to issues if multiple .jar files were present. The use of a specific JAR name via the environment variable eliminates this risk and makes the Docker build process more precise and reliable.