Expand FakeController and add User model and service
Enhanced the FakeController by adding new mapping functions, introduced UserService, and created a User model. This not only expands the functionality of the FakeController but also introduces the capability to manage users. Additionally, minor adjustments were made to gradle commands in the Dockerfile and to improve the reliability of the build process.